requirements.txt
hikari-crescent==1.0.0
hikari==2.0.0
mysql_connector_repackaged==0.3.1
openai==1.47.0
Requests==2.32.3
Unidecode==1.3.8
